fix: combine alert fixes 3235–3249 (#140)
Combines all 15 automated code scanning alert fixes for multiplication
result converted to larger type. These changes span
`include/stb_image.h`, `include/stb_image_write.h`, and
`src/preprocess.c` with consistent application of `size_t` casts to
prevent integer overflow before conversion to larger types. Overflow
checks are added where allocation sizes are stored back into `int`
variables such as `out_size`.

> [!NOTE]
> Merging this pull request will automatically close #125, #126, #127,
#128, #129, #130, #131, #132, #133, #134, #135, #136, #137, #138, #139

---------

+++ include/stb_image_write.h
@@ -888,8 +888,8 @@ static void *stbiw__sbgrowf(void **arr, int increment, int itemsize) {
   int m = *arr ? 2 * stbiw__sbm(*arr) + increment : increment + 1;
   void *p = STBIW_REALLOC_SIZED(
       *arr ? stbiw__sbraw(*arr) : 0,
-      *arr ? (stbiw__sbm(*arr) * itemsize + sizeof(int) * 2) : 0,
-      itemsize * m + sizeof(int) * 2);
+      *arr ? ((size_t)stbiw__sbm(*arr) * (size_t)itemsize + (size_t)sizeof(int) * 2) : 0,
+      (size_t)itemsize * (size_t)m + (size_t)sizeof(int) * 2);
   STBIW_ASSERT(p);
   if (p) {
     if (!*arr)
@@ -1211,7 +1211,7 @@ static void stbiw__encode_png_line(unsigned char *pixels, int stride_bytes,
       stbi__flip_vertically_on_write ? -stride_bytes : stride_bytes;
 
   if (type == 0) {
-    memcpy(line_buffer, z, width * n);
+    memcpy(line_buffer, z, (size_t)width * (size_t)n);
     return;
   }
 
@@ -1285,10 +1285,10 @@ STBIWDEF unsigned char *stbi_write_png_to_mem(const unsigned char *pixels,
     force_filter = -1;
   }
 
-  filt = (unsigned char *)STBIW_MALLOC((x * n + 1) * y);
+  filt = (unsigned char *)STBIW_MALLOC((((size_t)x * (size_t)n) + 1u) * (size_t)y);
   if (!filt)
     return 0;
-  line_buffer = (signed char *)STBIW_MALLOC(x * n);
+  line_buffer = (signed char *)STBIW_MALLOC((size_t)x * (size_t)n);
   if (!line_buffer) {
     STBIW_FREE(filt);
     return 0;
@@ -1326,7 +1326,7 @@ STBIWDEF unsigned char *stbi_write_png_to_mem(const unsigned char *pixels,
     // when we get here, filter_type contains the filter type, and line_buffer
     // contains the data
     filt[j * (x * n + 1)] = (unsigned char)filter_type;
-    STBIW_MEMMOVE(filt + j * (x * n + 1) + 1, line_buffer, x * n);
+    STBIW_MEMMOVE(filt + j * (x * n + 1) + 1, line_buffer, (size_t)x * (size_t)n);
   }
   STBIW_FREE(line_buffer);
   zlib = stbi_zlib_compress(filt, y * (x * n + 1), &zlen,
